+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v1.6.6 (unreleased)
+  - Display story summary at the top of the email story form
   - Added configuration option to set display order for 3rd party authentication modules
   - Fixed issue where help text did not display in configuration screen for items that were arrays
   - Ensure links in emailed stories are absolute